[PyTorch] Move main logic to unboxing.py
[ghstack-poisoned]
