+++ tools/codegen/api/cpp.py
@@ -305,81 +305,3 @@ def arguments(
             has_tensor_options=arguments.tensor_options is not None,
             cpp_no_default_args=cpp_no_default_args)
     ]
-
-
-def argumenttype_ivalue_convert_wrapper(arg: Argument, ival_str: str) -> (str, str):
-    result = argumenttype_ivalue_convert(arg.type, ival_str, arg.name)
-    return '\n'.join(result['code']), result['val_name']
-
-
-# Take an argument in JIT type format, returns the C++ code to convert an ivalue from stack to corresponding C++ type.
-def argumenttype_ivalue_convert(t: Type, ival: str, arg_name: str) -> Dict[str, Any]:
-    if isinstance(t, BaseType):
-        ctype = argumenttype_type(t=t, mutable=True, binds=name).cpp_type(strip_ref=True)
-        return {
-            "code": [f"{ctype} {arg_name}_base = {ival}.to<{ctype}>();"],
-            "val_name": f"{arg_name}_base",
-            "ctype": ctype,
-        }
-    elif isinstance(t, OptionalType):
-        in_name = arg_name + "_opt_in"
-        out_name = arg_name + "_opt_out"
-        connector = '\n\t'
-        res = argumenttype_ivalue_convert(t.elem, in_name, arg_name)
-        ctype = f"c10::optional<{res['ctype']}>"
-        code = f"""
-c10::optional<c10::IValue> {arg_name + "_opt"} = {ival}.toOptional<c10::IValue>();
-{ctype} {out_name};
-if ({arg_name + "_opt"}.has_value()) {{
-    const c10::IValue {in_name} = {arg_name + "_opt"}.value();
-    {connector.join(res['code'])}
-    {out_name} = {ctype}({res['val_name']});
-}} else {{
-    {out_name} = {ctype}();
-}}
-        """.split('\n')
-        return {
-            "code": code,
-            "val_name": out_name,
-            "ctype": ctype,
-        }
-    elif isinstance(t, ListType):
-        in_name = arg_name + "_list_in"
-        out_name = arg_name + "_list_out"
-        elem_name = arg_name + "_elem"
-        code = [f"const c10::List<c10::IValue> {in_name} = {ival}.toList();"]
-        res = argumenttype_ivalue_convert(t.elem, elem_name, arg_name)
-        connector = '\n\t'
-        # handle list type with size, e.g., bool[4]
-        if isinstance(t.elem, BaseType) and t.elem.name == BaseTy.bool and t.size:
-            ctype = f"std::array<{res['ctype']}, {t.size}>"
-            code.extend(f"""
-{ctype} {out_name} = as_array<{res['ctype']}, {t.size}>({in_name});
-            """.split('\n'))
-        # we have to use c10::List for optional element. e.g., Tensor?[] -> c10::List<c10::optional<at::Tensor>>
-        elif isinstance(t.elem, OptionalType):
-            ctype = f"c10::List<{res['ctype']}>"
-            code.extend(f"""
-{ctype} {out_name};
-for (c10::IValue {elem_name}: {in_name}) {{
-    {connector.join(res['code'])}
-    {out_name}.push_back({res['val_name']});
-}}
-            """.split('\n'))
-        else:
-            # use ArrayRef as default.
-            ctype = f"at::ArrayRef<{res['ctype']}>"
-            vec_name = arg_name + "_vec"
-            code.extend(f"""
-std::vector<{res['ctype']}> {vec_name};
-for (c10::IValue {elem_name}: {in_name}) {{
-    {connector.join(res['code'])}
-    {vec_name}.push_back({res['val_name']});
-}}
-{ctype} {out_name}({vec_name});
-            """.split('\n'))
-        return {
-            "code": code,
-            "val_name": out_name,
-            "ctype": ctype,
-        }
